run/vulnerability_detection/devign/requirements.txt
scikit-learn>=0.22.2
torch
torch-geometric
torch-sparse
torch-scatter
gensim==3.8.1
cpgclientlib==0.11.111
pandas>=1.0.1
